Get the root exception from the ServletException.
I'd say its a bug in ServletException not to take the rootCause into account when it generates
the stacktrace, though this code will workaround it. If there is a rootCause no one is interested
in the ServletException.
Now, e.g if it comes to an EvaluationException you'll be able (maybe by creating your own error.jsp)
to show the real exception. 


git-svn-id: https://svn.apache.org/repos/asf/myfaces/core/trunk@391944 13f79535-47bb-0310-9956-ffa450edef68