MYFACES-1681
thanks to Sun's Ryan Lubke, for pointing out, that MyFaces is too restrictive with respect
to null values (on UIComponent.getAttributes().put(...)).
The RI will only throw the NPE against a null value *if* there isn't an associated property. 
Myfaces does this as well, now.

JavaDoc is misleading, so Ryan will file a "Spec Bug" as well.

added fix for the problem
added test-case for this scenario.


git-svn-id: https://svn.apache.org/repos/asf/myfaces/core/branches/1_2_1@557580 13f79535-47bb-0310-9956-ffa450edef68