https://issues.apache.org/jira/browse/MYFACES-2112 https://issues.apache.org/jira/browse/MYFACES-2110
ongoing works in the ajax event handling, since the latest released sources of the ri are different to the spec information, the code has been made compatible to the latest released public ri sources! Also since the trinidad based logging facilities were suboptimal a decend javascript logging framework has been added, the code for backwards compatibility has been preserved so that  it still can be called from JSF2Utils (maybe changed later)

Next step implementation of the error handlers, and rechecking if we are still compatible to the RI in this regard!
(Trinidad might do different things in the event handling of the ajax core!)



git-svn-id: https://svn.apache.org/repos/asf/myfaces/core/branches/2_0_0@741158 13f79535-47bb-0310-9956-ffa450edef68