MYFACES-2290 - Add OSGi bundle information and bundle classloader / activator
committed David Jencks' patch for MYFACES-2290
=> Geronimo integration of myfaces 2 cannot proceed without this patch. 

The attached patch changes the osgi metadata installed by the maven-bundle-plugin so that myfaces can be used with osgi bundles with normal package versioning for javaee6, mostly allowing use of a servlet 3.0 spec jar. Note that this does not affect the ability to use myfaces 2 with earlier spec jars, it only allows myfaces to be used in ee6 containers that run under osgi. 

Thanks to David Jencks for the patch!

git-svn-id: https://svn.apache.org/repos/asf/myfaces/core/trunk@893759 13f79535-47bb-0310-9956-ffa450edef68