https://issues.apache.org/jira/browse/MYFACES-2747
Ok IE6 as usual hates dom manipulation while it renders the body and terminates upfront with an error (old ie6 bug never fixed and never will be), deferring is not an option since inline scripts might rely on
jsf.ajax already being present.
What I am doing now is simply to run the globalEval in case of IE6 instead of relying on the head manipulating loading routines.
I have changed the code my example now works again with IE6 even if it loads the jsf.js automatically within the body.

This problem occurs anyway only in debug mode when the subloading happens (new feature of the 2.0.1 codebase per request from Leonardo) and only on IE6 (and maybe ie7 but I doubt it) which is the only browser which has this problem.
In either way we have to add a fallback for those browsers that under no circumstances the scripts should be attached to the body but the head instead (while other browsers like chrome forbid explict html head alteration, one reason why globaleval for instance in jquery fails on webkit)


git-svn-id: https://svn.apache.org/repos/asf/myfaces/core/trunk@952920 13f79535-47bb-0310-9956-ffa450edef68