https://issues.apache.org/jira/browse/MYFACES-2849
Ok fixed, again I had a look at the jquery code, their method is one of the biggest hacks I have seen they append
a script element before the head on the document, and hope it gets parsed, this works only because firefox allows it, 
and not too many other browsers do auto eval anymore,
but I am not sure if this is not bound to fail on modern browsers. The window.onload method we do looks way saner to me, but it can get overwritten, for that case I have a body.onload fallback added in case the body already exists.
If someone does not chain but overwrites both, bad luck. I will check other frameworks, but I assume they chain onload handlers as well, I am definitely not going to introduce the jquery hack here which might break on some browsers (GlobalEval of jquery already does on chrome due to similar issues)



git-svn-id: https://svn.apache.org/repos/asf/myfaces/core/trunk@980706 13f79535-47bb-0310-9956-ffa450edef68