https://issues.apache.org/jira/browse/MYFACES-2876
reenabling winmobile 6.5 support, codebase, thanks to the overhauled delete code it works now better than ever.
The codebase is also working on opera 10, tested  also against blackberry 4.7.1, 5.x
and also works on webkit mobile which is basically all modern phones including blackberry 6.

Guess we now are done with the issue on the javascript side.

One sidenote, supporting IE Mobile 6.x is a real pain, since Micosoft is going to phase out this browser until the end of the year
I will probably stop supporting it sometime next year, depending on the pain it is to support it. I hope the next Mobile IE
will have a better experience, but given that it seems to be a version between ie7 and ie8 I doubt it will be very pleasant.




git-svn-id: https://svn.apache.org/repos/asf/myfaces/core/trunk@995434 13f79535-47bb-0310-9956-ffa450edef68